npr normalizer, split up home page server calls
