+++ server/api/homepagelatestnewsupdates.ts
@@ -1,0 +1,118 @@
+const config = useRuntimeConfig()
+import axios from 'axios'
+import humps from 'humps'
+import { cmsSources } from '~/composables/globals'
+import { estimateMp3Duration } from '~/server/utils/duration'
+
+// handleDuration is a helper function that checks if the estimated duration is available and if not, it estimates it using the audio URL in the estimateMp3Duration function.
+const handleDuration = async (estimatedDuration: number, audioURL: string) => {
+	if (!estimatedDuration || typeof estimatedDuration !== 'number' || estimatedDuration === 0) {
+		return await estimateMp3Duration(audioURL);
+	} else {
+		return estimatedDuration
+	}
+}
+
+const getLocalNewscast = async () => {
+	try {
+		const options = {
+			method: 'GET',
+			url: `${config.public.PUBLISHER_BASE_API}v3/story/latest-newscast/`,
+		};
+		const res = await axios(options);
+		const resData = humps.camelizeKeys(res.data).data;
+		resData.attributes.file = resData.attributes.audio;
+		resData.attributes.image = resData.attributes.headers.brand.logoImage.template;
+		resData.attributes.duration = await handleDuration(resData.attributes.estimatedDuration, resData.attributes.audio);
+		resData.attributes.cardTitle = 'NYC Headlines';
+		resData.attributes.showTitle = resData.attributes.channelTitle;
+		resData.attributes.type = resData.type;
+		resData.attributes.id = resData.id;
+		resData.attributes.cmsSource = cmsSources.PUBLISHER;
+		resData.attributes.hideFavorite = true;
+		return resData.attributes;
+	} catch (e) {
+		////console.log(e);
+	}
+	return null
+}
+
+const getNationalNewscast = async () => {
+	try {
+		const options = {
+			method: 'GET',
+			url: `${config.public.PUBLISHER_BASE_API}v3/story/npr-newscast`,
+		};
+		const res = await axios(options);
+		const resData = humps.camelizeKeys(res.data).data;
+		resData.attributes.file = resData.attributes.audio;
+		resData.attributes.image = 'https://media.wnyc.org/i/%s/%s/%s/%s/2023/09/npr-news-now.jpeg';
+		resData.attributes.duration = await handleDuration(resData.attributes.estimatedDuration, resData.attributes.audio);
+		resData.attributes.cardTitle = 'NPR News Now';
+		resData.attributes.showTitle = resData.attributes.channelTitle;
+		resData.attributes.type = resData.type;
+		resData.attributes.id = resData.id;
+		resData.attributes.cmsSource = cmsSources.PUBLISHER;
+		resData.attributes.hideFavorite = true;
+		return resData.attributes;
+	} catch (e) {
+		////console.log(e);
+	}
+	return null
+}
+
+// Get NYC-NOW newscast from the WNYC API
+const getNYCNowNewscast = async () => {
+	try {
+		const options = {
+			method: 'GET',
+			url: `${config.public.PUBLISHER_BASE_API}v3/story/`,
+			params: {
+				'channel': 'nyc-now-podcast',
+				'ordering': '-newsdate',
+				'page_size': '1',
+			},
+		};
+		const res = await axios(options);
+		const resData = humps.camelizeKeys(res.data).data[0];
+		resData.attributes.file = resData.attributes.audio;
+		resData.attributes.image = 'https://media.wnyc.org/i/%s/%s/%s/%s/2023/04/NYNOW_WNYC_LOGO_HEX_1400PX.png';
+		resData.attributes.duration = await handleDuration(resData.attributes.estimatedDuration, resData.attributes.audio);
+		resData.attributes.cardTitle = 'NYC Now';
+		resData.attributes.showTitle = resData.attributes.channelTitle;
+		resData.attributes.type = resData.type;
+		resData.attributes.id = resData.id;
+		resData.attributes.cmsSource = cmsSources.PUBLISHER;
+		resData.attributes.hideFavorite = true;
+		return resData.attributes;
+	} catch (e) {
+		//console.log(e);
+	}
+	return null
+}
+
+/**
+ * Compress and simplify the global nav data.
+ * Reachable /api/homepage
+ */
+export default defineEventHandler(async (event) => {
+	let res = event?.node?.res;
+	// WNYC NOW Newscast is only available on weekdays between 7am and 7pm
+	// If it is not available, use the local newscast instead.
+	const requestTime = new Date();
+	const day = requestTime.getDay();
+	const hour = requestTime.getHours();
+	let local_newscast;
+	if (day > 0 && day < 6 && hour > 6 && hour < 19) {
+		local_newscast = await getNYCNowNewscast();
+	} else {
+		local_newscast = await getLocalNewscast();
+	}
+	const national_newscast = await getNationalNewscast();
+	res.setHeader('Cache-Control', 'maxage=900, stale-while-revalidate');
+
+	return {
+		local_newscast: local_newscast,
+		national_newscast: national_newscast,
+	}
+})
